cu29::bevy_reflect

Function set_apply

pub fn set_apply<M>(a: &mut M, b: &(dyn PartialReflect + 'static))
where
    M: Set,

Applies the elements of reflected set b to the corresponding elements of set a.

If a value from b does not exist in a, the value is cloned and inserted. If a value from a does not exist in b, the value is removed.

Panics

This function panics if b is not a reflected set.