cu29::bevy_reflect::prelude

Trait GetTupleStructField

pub trait GetTupleStructField {
    // Required methods
    fn get_field<T>(&self, index: usize) -> Option<&T>
       where T: Reflect;
    fn get_field_mut<T>(&mut self, index: usize) -> Option<&mut T>
       where T: Reflect;
}

A convenience trait which combines fetching and downcasting of tuple struct fields.

Example

use bevy_reflect::{GetTupleStructField, Reflect};

#[derive(Reflect)]
struct Foo(String);

let mut foo = Foo("Hello, world!".to_string());

foo.get_field_mut::<String>(0).unwrap().truncate(5);
assert_eq!(foo.get_field::<String>(0), Some(&"Hello".to_string()));

Required Methods

fn get_field<T>(&self, index: usize) -> Option<&T>
where T: Reflect,

Returns a reference to the value of the field with index index, downcast to T.

fn get_field_mut<T>(&mut self, index: usize) -> Option<&mut T>
where T: Reflect,

Returns a mutable reference to the value of the field with index index, downcast to T.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.