cu29::bevy_reflect

Module set

A trait used to power set-like operations via reflection.

Structs

DynamicSet

An unordered set of reflected values.

SetInfo

A container for compile-time set info.

Traits

Set

A trait used to power set-like operations via reflection.

Functions

set_apply

Applies the elements of reflected set b to the corresponding elements of set a.

set_debug

The default debug formatter for Set types.

set_partial_eq

Compares a Set with a PartialReflect value.

set_try_apply

Tries to apply the elements of reflected set b to the corresponding elements of set a and returns a Result.