cu29::bevy_reflect

Module structs

Traits and types used to power struct-like operations via reflection.

Structs

DynamicStruct

A struct type which allows fields to be added at runtime.

FieldIter

An iterator over the names and fields of a struct.

StructInfo

A container for compile-time named struct info.

Traits

GetField

A convenience trait which combines fetching and downcasting of struct fields.

Struct

A trait used to power struct-like operations via reflection.

Functions

struct_debug

The default debug formatter for Struct types.

struct_partial_cmp

Lexicographically compares two Struct values and returns their ordering.

struct_partial_eq

Compares a Struct with a PartialReflect value.