cu29
Module config

This module defines the configuration of the copper runtime. The configuration is a directed graph where nodes are tasks and edges are connections between tasks. The configuration is serialized in the RON format. The configuration is used to generate the runtime code at compile time.

Structs

Cnx

This represents a connection between 2 tasks (nodes) in the configuration graph.

ComponentConfig

This is the configuration of a component (like a task config or a monitoring config):w It is a map of key-value pairs. It is given to the new method of the task implementation.

CuConfig

CuConfig is the programmatic representation of the configuration graph. It is a directed graph where nodes are tasks and edges are connections between tasks.

MonitorConfig

Node

A node in the configuration graph. A node represents a Task in the system Graph.

Value

Wrapper around the ron::Value to allow for custom serialization.

Functions

read_configuration

Read a copper configuration from a file.

Type Aliases

NodeId

NodeId is the unique identifier of a node in the configuration graph for petgraph and the code generation.