cu29::config

Struct CuConfig

pub struct CuConfig {
    pub monitor: Option<MonitorConfig>,
    pub logging: Option<LoggingConfig>,
    pub runtime: Option<RuntimeConfig>,
    pub resources: Vec<ResourceBundleConfig>,
    pub bridges: Vec<BridgeConfig>,
    pub graphs: ConfigGraphs,
}

CuConfig is the programmatic representation of the configuration graph. It is a directed graph where nodes are tasks and edges are connections between tasks.

The core of CuConfig is its graphs field which can be either a simple graph or a collection of mission-specific graphs. The graph structure is based on petgraph.

Fields
monitor: Option<MonitorConfig>

Optional monitoring configuration

logging: Option<LoggingConfig>

Optional logging configuration

runtime: Option<RuntimeConfig>

Optional runtime configuration

resources: Vec<ResourceBundleConfig>

Declarative resource bundle definitions

bridges: Vec<BridgeConfig>

Declarative bridge definitions that are yet to be expanded into the graph

graphs: ConfigGraphs

Graph structure - either a single graph or multiple mission-specific graphs

Implementations

impl CuConfig

The implementation has a lot of convenience methods to manipulate the configuration to give some flexibility into programmatically creating the configuration.

pub fn new_simple_type() -> CuConfig

pub fn new_mission_type() -> CuConfig

pub fn serialize_ron(&self) -> Result<String, CuError>

pub fn deserialize_ron(ron: &str) -> Result<CuConfig, CuError>

pub fn render( &self, output: &mut dyn Write, mission_id: Option<&str>, ) -> Result<(), CuError>

Render the configuration graph in the dot format.

pub fn get_all_instances_configs( &self, mission_id: Option<&str>, ) -> Vec<Option<&ComponentConfig>>

pub fn get_graph(&self, mission_id: Option<&str>) -> Result<&CuGraph, CuError>

pub fn get_graph_mut( &mut self, mission_id: Option<&str>, ) -> Result<&mut CuGraph, CuError>

pub fn get_monitor_config(&self) -> Option<&MonitorConfig>

pub fn get_runtime_config(&self) -> Option<&RuntimeConfig>

pub fn validate_logging_config(&self) -> Result<(), CuError>

Validate the logging configuration to ensure section pre-allocation sizes do not exceed slab sizes. This method is wrapper around LoggingConfig::validate