Module cu29::copperlist

CopperList is the main data structure used by Copper to communicate between tasks. It is a queue that can be used to store preallocated messages between tasks in memory order.

Structs

CopperLiskMask

Not implemented yet. This mask will be used to for example filter out necessary regions of a copper list between remote systems.

CopperList

CuListsManager

This structure maintains the entire memory needed by Copper for one loop for the inter tasks communication within a process. P or Payload is typically a Tuple of various types of messages that are exchanged between tasks. N is the maximum number of in flight Copper List the runtime can support.

Enums

CopperListState

Those are the possible states along the lifetime of a CopperList.

Type Aliases

AscIter

AscIterMut

Iter

IterMut