cu29
Module cutask

This module contains all the main definition of the traits you need to implement or interact with to create a Copper task.

Structs

CuCompactString

CuMsg

CuMsg is the envelope holding the msg payload and the metadata between tasks.

CuMsgMetadata

CuMsgMetadata is a structure that contains metadata common to all CuMsgs.

Traits

CuMsgPack

CuMsgPayload

CuSinkTask

A Sink Task is a task that only consumes messages. For example drivers for actuators are Sink Tasks.

CuSrcTask

A Src Task is a task that only produces messages. For example drivers for sensors are Src Tasks. They are in push mode from the runtime. To set the frequency of the pulls and align them to any hw, see the runtime configuration.

CuTask

This is the most generic Task of copper. It is a “transform” task deriving an output from an input.

CuTaskLifecycle

The CuTaskLifecycle trait is the base trait for all tasks in Copper. It defines the lifecycle of a task. It provides a default empty implementation as all those execution steps are optional.

Freezable

The internal state of a task needs to be serializable so the framework can take a snapshop of the task graph.