Trait cu29::cutask::CuTask

pub trait CuTask: CuTaskLifecycle {
    type Input: CuMsgPayload;
    type Output: CuMsgPayload;

    // Required method
    fn process(
        &mut self,
        clock: &RobotClock,
        input: &CuMsg<Self::Input>,
        output: &mut CuMsg<Self::Output>,
    ) -> CuResult<()>;
}

This is the most generic Task of copper. It is a “transform” task deriving an output from an input.

Required Associated Types

type Input: CuMsgPayload

type Output: CuMsgPayload

Required Methods

fn process( &mut self, clock: &RobotClock, input: &CuMsg<Self::Input>, output: &mut CuMsg<Self::Output>, ) -> CuResult<()>

Process is the most critical execution of the task. The goal will be to produce the output message as soon as possible. Use preprocess to prepare the task to make this method as short as possible.

Object Safety

This trait is not object safe.