cu29::cutask
Trait CuTaskLifecycle

pub trait CuTaskLifecycle: Freezable {
    // Required method
    fn new(config: Option<&ComponentConfig>) -> CuResult<Self>
       where Self: Sized;

    // Provided methods
    fn start(&mut self, _clock: &RobotClock) -> CuResult<()> { ... }
    fn preprocess(&mut self, _clock: &RobotClock) -> CuResult<()> { ... }
    fn postprocess(&mut self, _clock: &RobotClock) -> CuResult<()> { ... }
    fn stop(&mut self, _clock: &RobotClock) -> CuResult<()> { ... }
}

The CuTaskLifecycle trait is the base trait for all tasks in Copper. It defines the lifecycle of a task. It provides a default empty implementation as all those execution steps are optional.

Required Methods

fn new(config: Option<&ComponentConfig>) -> CuResult<Self>
where Self: Sized,

Provided Methods

fn start(&mut self, _clock: &RobotClock) -> CuResult<()>

Start is called once for a long period of time. Here you need to initialize everything your task will need for the duration of its lifetime.

fn preprocess(&mut self, _clock: &RobotClock) -> CuResult<()>

This is a method called by the runtime before “process”. This is a kind of best effort, as soon as possible call to give a chance for the task to do some work before to prepare to make “process” as short as possible.

fn postprocess(&mut self, _clock: &RobotClock) -> CuResult<()>

This is a method called by the runtime after “process”. It is best effort a chance for the task to update some state after process is out of the way. It can be use for example to maintain statistics etc. that are not time critical for the robot.

fn stop(&mut self, _clock: &RobotClock) -> CuResult<()>

Call at the end of the lifecycle of the task.

Object Safety

This trait is not object safe.