cu29::prelude::bevy_reflect

Module map

Traits and types used to power map-like operations via reflection.

Structs

DynamicMap

An unordered mapping between reflected values.

MapInfo

A container for compile-time map info.

Traits

Map

A trait used to power map-like operations via reflection.

Functions

map_apply

Applies the elements of reflected map b to the corresponding elements of map a.

map_debug

The default debug formatter for Map types.

map_partial_cmp

Lexicographically compares two Map values according to their iteration order (suitable for ordered maps like BTreeMap).

map_partial_eq

Compares a Map with a PartialReflect value.

map_try_apply

Tries to apply the elements of reflected map b to the corresponding elements of map a and returns a Result.