cu29::prelude

Module config

This module defines the configuration of the copper runtime. The configuration is a directed graph where nodes are tasks and edges are connections between tasks. The configuration is serialized in the RON format. The configuration is used to generate the runtime code at compile time.

Structs

BridgeConfig

Declarative definition of a bridge component with a list of channels.

Cnx

This represents a connection between 2 tasks (nodes) in the configuration graph.

ComponentConfig

This is the configuration of a component (like a task config or a monitoring config):w It is a map of key-value pairs. It is given to the new method of the task implementation.

ConfigError

CuConfig

CuConfig is the programmatic representation of the configuration graph. It is a directed graph where nodes are tasks and edges are connections between tasks.

CuGraph

IncludesConfig

Includes are used to include other configuration files.

LoggingConfig

MissionsConfig

Missions are used to generate alternative DAGs within the same configuration.

MonitorConfig

Node

A node in the configuration graph. A node represents a Task in the system Graph.

NodeLogging

Configuration for logging in the node.

ResourceBundleConfig

Declarative definition of a resource bundle.

RuntimeConfig

Value

Wrapper around the ron::Value to allow for custom serialization.

Enums

BridgeChannelConfigRepresentation

Directional mapping for bridge channels.

ConfigGraphs

CuDirection

A simple wrapper enum for petgraph::Direction, designed to be converted into it via the From trait.

Flavor

Distinguishes regular tasks from bridge nodes so downstream stages can apply bridge-specific instantiation rules.

Constants

DEFAULT_KEYFRAME_INTERVAL

Functions

read_configuration

Read a copper configuration from a file.

read_configuration_str

Type Aliases

NodeId

NodeId is the unique identifier of a node in the configuration graph for petgraph and the code generation.