cu29::prelude::config
Struct CuConfig

pub struct CuConfig {
    pub graph: StableGraph<Node, Cnx>,
    pub monitor: Option<MonitorConfig>,
    pub logging: Option<LoggingConfig>,
}

CuConfig is the programmatic representation of the configuration graph. It is a directed graph where nodes are tasks and edges are connections between tasks.

Fields
graph: StableGraph<Node, Cnx>monitor: Option<MonitorConfig>logging: Option<LoggingConfig>
Implementations

impl CuConfig

The implementation has a lot of convenience methods to manipulate the configuration to give some flexibility into programmatically creating the configuration.

pub fn add_node(&mut self, node: Node) -> u32

Add a new node to the configuration graph.

pub fn get_node(&self, node_id: u32) -> Option<&Node>

Get the node with the given id.

pub fn get_node_mut(&mut self, node_id: u32) -> Option<&mut Node>

Get the node with the given id mutably.

pub fn get_node_output_msg_type(&self, node_id: &str) -> Option<String>

this is more like infer from the connections of this node.

pub fn get_node_input_msg_type(&self, node_id: &str) -> Option<String>

this is more like infer from the connections of this node.

pub fn get_src_edges(&self, node_id: u32) -> Vec<usize>

Get the list of edges that are connected to the given node as a source.

pub fn get_dst_edges(&self, node_id: u32) -> Vec<usize>

Get the list of edges that are connected to the given node as a destination.

pub fn get_edge_weight(&self, index: usize) -> Option<Cnx>

pub fn get_all_nodes(&self) -> Vec<(u32, &Node)>

Convenience method to get all nodes in the configuration graph.

pub fn connect_ext( &mut self, source: u32, target: u32, msg_type: &str, batch: Option<u32>, store: Option<bool>, )

Adds an edge between two nodes/tasks in the configuration graph. msg_type is the type of message exchanged between the two nodes/tasks. batch is the number of messages to batch before sending the buffer. store tells Copper if it needs to log the messages.

pub fn connect(&mut self, source: u32, target: u32, msg_type: &str)

Adds an edge between two nodes/tasks in the configuration graph. msg_type is the type of message exchanged between the two nodes/tasks.

pub fn serialize_ron(&self) -> String

pub fn deserialize_ron(ron: &str) -> CuConfig

pub fn render(&self, output: &mut dyn Write)

Render the configuration graph in the dot format.

pub fn get_all_instances_configs(&self) -> Vec<Option<&ComponentConfig>>

pub fn get_monitor_config(&self) -> Option<&MonitorConfig>

pub fn validate_logging_config(&self) -> Result<(), CuError>

Validate the logging configuration to ensure section pre-allocation sizes do not exceed slab sizes. This method is wrapper around LoggingConfig::validate