cu29::prelude

Module memmap

This is the memory map file implementation for the unified logger for Copper. It is std only.

Structs

MmapSectionStorage

MmapUnifiedLoggerBuilder

Use this builder to create a new DataLogger.

MmapUnifiedLoggerRead

A read side of the memory map based unified logger.

MmapUnifiedLoggerWrite

A write side of the datalogger.

UnifiedLoggerIOReader

This a convenience wrapper around the UnifiedLoggerRead to implement the Read trait.

Enums

MmapUnifiedLogger

Holds the read or write side of the datalogger.