cu29::prelude::monitoring

Trait CuPayloadSize

pub trait CuPayloadSize {
    // Provided methods
    fn raw_bytes(&self) -> usize { ... }
    fn handle_bytes(&self) -> usize { ... }
}

Lightweight trait to estimate the amount of data a payload will contribute when serialized. Default implementations return the stack size; specific types override to report dynamic data.

Provided Methods

fn raw_bytes(&self) -> usize

Total bytes represented by the payload in memory (stack + heap backing).

fn handle_bytes(&self) -> usize

Bytes that correspond to reusable/pooled handles (used for IO budgeting).