cu29::prelude

Module pool

Structs

AlignedBuffer

A buffer that is aligned to a specific size with the Element of type E.

CuHandle

A shareable handle to an Array coming from a pool (either host or device).

CuHostMemoryPool

A pool of host memory buffers.

Enums

CuHandleInner

A Handle to a Buffer. For onboard usages, the buffer should be Pooled (ie, coming from a preallocated pool). The Detached version is for offline usages where we don’t really need a pool to deserialize them.

Traits

ArrayLike

CuPool

A CuPool is a pool of buffers that can be shared between different parts of the code. Handles can be stored locally in the tasks and shared between them.

DeviceCuPool

A device memory pool can copy data from a device to a host memory pool on top.

ElementType

Basic Type that can be used in a buffer in a CuPool.

PoolMonitor

Trait for a Pool to exposed to be monitored by the monitoring API.

Functions

pools_statistics

Get the list of pools and their statistics. We use SmallVec here to avoid heap allocations while the stack is running.