cu29::prelude::pool

Trait PoolMonitor

pub trait PoolMonitor: Send + Sync {
    // Required methods
    fn id(&self) -> ArrayString<64>;
    fn space_left(&self) -> usize;
    fn total_size(&self) -> usize;
    fn buffer_size(&self) -> usize;
}

Trait for a Pool to exposed to be monitored by the monitoring API.

Required Methods

fn id(&self) -> ArrayString<64>

A unique and descriptive identifier for the pool.

fn space_left(&self) -> usize

Number of buffer slots left in the pool.

fn total_size(&self) -> usize

Total size of the pool in number of buffers.

fn buffer_size(&self) -> usize

Size of one buffer