cu29::prelude::reflect

Module prelude

The reflect prelude.

This includes the most common types in this crate, re-exported for your convenience.

Structs

ReflectDefault

A struct used to provide the default value of a type.

ReflectDeserialize

A struct used to deserialize reflected instances of a type.

ReflectFromReflect

Type data that represents the FromReflect trait and allows it to be used dynamically.

ReflectSerialize

A struct used to serialize reflected instances of a type.

Traits

FromReflect

A trait that enables types to be dynamically constructed from reflected data.

GetField

A convenience trait which combines fetching and downcasting of struct fields.

GetPath

A trait which allows nested Reflect values to be retrieved with path strings.

GetTupleStructField

A convenience trait which combines fetching and downcasting of tuple struct fields.

PartialReflect

The foundational trait of bevy_reflect, used for accessing and modifying data dynamically.

Reflect

A core trait of bevy_reflect, used for downcasting to concrete types.

ReflectPath

Something that can be interpreted as a reflection path in GetPath.

Struct

A trait used to power struct-like operations via reflection.

TupleStruct

A trait used to power tuple struct-like operations via reflection.

TypePath

A static accessor to type paths and names.

Attribute Macros

reflect_trait

A macro that automatically generates type data for traits, which their implementors can then register.

Derive Macros

FromReflect

Derives the FromReflect trait.

Reflect

The main derive macro used by bevy_reflect for deriving its Reflect trait.

TypePath

Derives the TypePath trait, providing a stable alternative to [std::any::type_name].