cu29::prelude::reflect

Trait TypeData

pub trait TypeData:
    Downcast
    + Send
    + Sync {
    // Required method
    fn clone_type_data(&self) -> Box<dyn TypeData>;
}

A trait used to type-erase type metadata.

Type data can be registered to the TypeRegistry and stored on a type’s TypeRegistration.

While type data is often generated using the #[reflect_trait] macro, almost any type that implements Clone can be considered “type data”. This is because it has a blanket implementation over all T where T: Clone + Send + Sync + 'static.

See the crate-level documentation for more information on type data and type registration.

Required Methods

fn clone_type_data(&self) -> Box<dyn TypeData>

Creates a type-erased clone of this value.

Implementations

impl dyn TypeData

pub fn is<__T>(&self) -> bool
where __T: TypeData,

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T>( self: Box<dyn TypeData>, ) -> Result<Box<__T>, Box<dyn TypeData>>
where __T: TypeData,

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn’t.

pub fn downcast_rc<__T>( self: Rc<dyn TypeData>, ) -> Result<Rc<__T>, Rc<dyn TypeData>>
where __T: TypeData,

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn’t.

pub fn downcast_ref<__T>(&self) -> Option<&__T>
where __T: TypeData,

Returns a reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_mut<__T>(&mut self) -> Option<&mut __T>
where __T: TypeData,

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn’t.