cu29::reflect

Module list

Traits and types used to power list-like operations via reflection.

Structs

DynamicList

A list of reflected values.

ListInfo

A container for compile-time list info.

ListIter

An iterator over an List.

Traits

List

A trait used to power list-like operations via reflection.

Functions

list_apply

Applies the elements of b to the corresponding elements of a.

list_debug

The default debug formatter for List types.

list_hash

Returns the u64 hash of the given list.

list_partial_cmp

Lexicographically compares two List values and returns their ordering.

list_partial_eq

Compares a List with a Reflect value.

list_try_apply

Tries to apply the elements of b to the corresponding elements of a and returns a Result.