cu29::reflect

Module tuple

Traits and types used to power tuple-like operations via reflection.

Structs

DynamicTuple

A tuple which allows fields to be added at runtime.

TupleFieldIter

An iterator over the field values of a tuple.

TupleInfo

A container for compile-time tuple info.

Traits

GetTupleField

A convenience trait which combines fetching and downcasting of tuple fields.

Tuple

A trait used to power tuple-like operations via reflection.

Functions

tuple_apply

Applies the elements of b to the corresponding elements of a.

tuple_debug

The default debug formatter for Tuple types.

tuple_partial_cmp

Lexicographically compares two Tuple values and returns their ordering.

tuple_partial_eq

Compares a Tuple with a PartialReflect value.

tuple_try_apply

Tries to apply the elements of b to the corresponding elements of a and returns a Result.