cu29::resource

Trait ResourceBundle

pub trait ResourceBundle: Sized + ResourceBundleDecl {
    // Required method
    fn build(
        bundle: BundleContext<Self>,
        config: Option<&ComponentConfig>,
        manager: &mut ResourceManager,
    ) -> Result<(), CuError>;
}

Bundle providers implement this trait to populate the ResourceManager with concrete resources for a given bundle id.

Required Methods

fn build( bundle: BundleContext<Self>, config: Option<&ComponentConfig>, manager: &mut ResourceManager, ) -> Result<(), CuError>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.