cu29::simulation

Struct CuSimSinkTask

pub struct CuSimSinkTask<I> { /* private fields */ }

This is a placeholder task for a sink task for the simulations. It basically does nothing in place of a real driver so it won’t try to initialize any hardware.