Crate cu29

Re-exports

pub use config::read_configuration;

pub use cu29_clock as clock;

Modules

config

This module defines the configuration of the copper runtime. The configuration is a directed graph where nodes are tasks and edges are connections between tasks. The configuration is serialized in the RON format. The configuration is used to generate the runtime code at compile time.

copperlist

CopperList is the main data structure used by Copper to communicate between tasks. It is a queue that can be used to store preallocated messages between tasks in memory order.

curuntime

CuRuntime is the heart of what copper is running on the robot. It is exposed to the user via the copper_runtime macro injecting it as a field in their application struct.

cutask

This module contains all the main definition of the traits you need to implement or interact with to create a Copper task.

monitoring

Some basic internal monitoring tooling Copper uses to monitor itself and the tasks it is running.

Structs

CuError

Common copper Error type.

Enums

UnifiedLogType

Defines the types of what can be logged in the unified logger.

Traits

CopperListPayload

A CopperListPayload needs to be encodable, decodable and fixed size in memory.

WriteStream

Defines a basic write, append only stream trait to be able to log or send serializable objects.

Type Aliases

CuResult