Trait cu29::cutask::CuSinkTask

pub trait CuSinkTask: CuTaskLifecycle {
    type Input: CuMsgPayload;

    // Required method
    fn process(
        &mut self,
        clock: &RobotClock,
        input: &mut CuMsg<Self::Input>,
    ) -> CuResult<()>;
}

A Sink Task is a task that only consumes messages. For example drivers for actuators are Sink Tasks.

Required Associated Types

type Input: CuMsgPayload

Required Methods

fn process( &mut self, clock: &RobotClock, input: &mut CuMsg<Self::Input>, ) -> CuResult<()>

Process is the most critical execution of the task. The goal will be to produce the output message as soon as possible. Use preprocess to prepare the task to make this method as short as possible. TODO: Still on the fence for this mutable input but sometimes you want to record one last thing before logging the message.

Object Safety

This trait is not object safe.