cu29

Module monitoring

Some basic internal monitoring tooling Copper uses to monitor itself and the tasks it is running.

Structs

CopperListInfo

CopperListIoStats

Reported data about CopperList IO for a single iteration.

CountingAlloc

A simple allocator that counts the number of bytes allocated and deallocated.

CuDurationStatistics

A Specialized statistics object for CuDuration. It will also keep track of the jitter between the values.

LiveStatistics

Accumulative stat object that can give your some real time statistics. Uses a fixed-size bucketed histogram for accurate percentile calculations.

MonitorConnection

MonitorNode

MonitorTopology

NoMonitor

A do nothing monitor if no monitor is provided. This is basically defining the default behavior of Copper in case of error.

Enums

ComponentKind

CuTaskState

The state of a task.

Decision

Monitor decision to be taken when a task errored out.

Traits

CuMonitor

Trait to implement a monitoring task.

CuPayloadSize

Lightweight trait to estimate the amount of data a payload will contribute when serialized. Default implementations return the stack size; specific types override to report dynamic data.

Functions

build_monitor_topology

Derive a monitor-friendly topology from the runtime configuration.