cu29::prelude

Struct CuRuntime

pub struct CuRuntime<CT, CB, P, M, const NBCL: usize>
where
    P: CopperListTuple,
    M: CuMonitor,
{
    pub clock: RobotClock,
    pub tasks: CT,
    pub bridges: CB,
    pub resources: ResourceManager,
    pub monitor: M,
    pub copperlists_manager: CopperListsManager<P, NBCL>,
    pub keyframes_manager: KeyFramesManager,
    pub runtime_config: RuntimeConfig,
}

This is the main structure that will be injected as a member of the Application struct. CT is the tuple of all the tasks in order of execution. CL is the type of the copper list, representing the input/output messages for all the tasks.

Fields
clock: RobotClock

The base clock the runtime will be using to record time.

tasks: CT

The tuple of all the tasks in order of execution.

bridges: CB

Tuple of all instantiated bridges.

resources: ResourceManager

Resource registry kept alive for tasks borrowing shared handles.

monitor: M

The runtime monitoring.

copperlists_manager: CopperListsManager<P, NBCL>

The logger for the copper lists (messages between tasks)

keyframes_manager: KeyFramesManager

The logger for the state of the tasks (frozen tasks)

runtime_config: RuntimeConfig

The runtime configuration controlling the behavior of the run loop

Implementations

impl<CT, CB, P, M, const NBCL: usize> CuRuntime<CT, CB, P, M, NBCL>
where P: CopperListTuple + CuListZeroedInit + Default + 'static, M: CuMonitor,

pub fn new( clock: RobotClock, config: &CuConfig, mission: Option<&str>, resources_instanciator: impl Fn(&CuConfig) -> Result<ResourceManager, CuError>, tasks_instanciator: impl for<'c> Fn(Vec<Option<&'c ComponentConfig>>, &mut ResourceManager) -> Result<CT, CuError>, monitor_instanciator: impl Fn(&CuConfig) -> M, bridges_instanciator: impl Fn(&CuConfig, &mut ResourceManager) -> Result<CB, CuError>, copperlists_logger: impl WriteStream<CopperList<P>> + 'static, keyframes_logger: impl WriteStream<KeyFrame> + 'static, ) -> Result<CuRuntime<CT, CB, P, M, NBCL>, CuError>

pub fn new_with_resources( clock: RobotClock, config: &CuConfig, mission: Option<&str>, resources: ResourceManager, tasks_instanciator: impl for<'c> Fn(Vec<Option<&'c ComponentConfig>>, &mut ResourceManager) -> Result<CT, CuError>, monitor_instanciator: impl Fn(&CuConfig) -> M, bridges_instanciator: impl Fn(&CuConfig, &mut ResourceManager) -> Result<CB, CuError>, copperlists_logger: impl WriteStream<CopperList<P>> + 'static, keyframes_logger: impl WriteStream<KeyFrame> + 'static, ) -> Result<CuRuntime<CT, CB, P, M, NBCL>, CuError>